#251. Migrate filters to epoxy.
