+++ app/src/main/java/com/bogdan/codeforceswatcher/features/filters/models/FilterItem.kt
@@ -1,0 +1,8 @@
+package com.bogdan.codeforceswatcher.features.filters.models
+
+data class FilterItem(
+    val imageId: Int?,
+    val title: String,
+    val isChecked: Boolean,
+    val onSwitchTap: (Boolean) -> Unit
+)
